requirements.txt
streamlit
PyPDF2
langchain
langchain-openai
python-dotenv